Backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
exceptiongroup==1.1.3
fastapi==0.103.2
greenlet==3.0.0
h11==0.14.0
httptools==0.6.0
idna==3.4
jwt==1.3.1
numpy==1.26.0
pandas==2.1.1
psycopg2==2.9.9
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
